tracks/python/requirements.txt
# Core Dependencies
deepgram-sdk>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
python-Levenshtein>=0.20.0
python-dotenv>=1.0.0

# Async Support
aiohttp>=3.9.0
httpx>=0.25.0
aiosqlite>=0.19.0
asyncio-throttle>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: Visualization and Monitoring
matplotlib>=3.7.0
plotly>=5.17.0
dash>=2.14.0
prometheus-client>=0.19.0

# Utilities
tqdm>=4.66.0  # Progress bars
tabulate>=0.9.0  # Pretty tables
rich>=13.7.0  # Rich terminal output